Modifed file structure
